services/rec_svc/requirements.txt
transformers==4.44.2
huggingface_hub<0.26  
fastapi
uvicorn[standard]
torch
torchvision
transformers
faiss-cpu
Pillow
numpy

